backend/statistics/requirements.txt
fastapi
uvicorn[standard]
motor
databases[asyncpg]
python-dotenv
pydantic
bson
